+++ lib/node.dart
@@ -2,7 +2,7 @@ import 'dart:js_interop';
 
 /// Node exports
 @JS()
-external JSObject exports;
+external JSObject get exports;
 
 /// Node require function
 @JS()
@@ -13,3 +13,13 @@ external JSObject require(String name);
 extension type Agent._(JSObject _) implements JSObject {
   // TODO: Add members
 }
+
+/// Node process object
+extension type Process._(JSObject _) implements JSObject {
+  /// Process environment
+  external JSObject get env;
+}
+
+/// Access to the node process object
+@JS()
+external Process get process;